tree clean

+++ src/main/java/com/codeup/springblog/Controllers/PostController.java
@@ -1,28 +1,28 @@
 package com.codeup.springblog.Controllers;
 
 import com.codeup.springblog.Models.Post;
+import com.codeup.springblog.Models.User;
 import com.codeup.springblog.Repository.PostRepository;
 import com.codeup.springblog.Repository.UserRepository;
 import com.codeup.springblog.Services.EmailService;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
-
+import java.security.Principal;
 
 
 @Controller
 public class PostController {
 
     private PostRepository postDao;
-    private final UserRepository userDao;
-
-    @Autowired
+    private UserRepository userDao;
     private EmailService emailService;
 
-    public PostController(PostRepository postDao, UserRepository userDao ){
+    public PostController(PostRepository postDao, UserRepository userDao, EmailService emailService){
         this.postDao = postDao;
         this.userDao = userDao;
+        this.emailService = emailService;
     }
 
     @GetMapping("/posts")
@@ -33,39 +33,39 @@ public class PostController {
 
 
     @GetMapping("/posts/{id}")
-    public String getPost(@PathVariable long id, Model model){
-//        Post post1 = new Post( id, "Europa's First Post", "Remote Learning Today!");
-//        model.addAttribute("title", post1.getTitle());
-//        model.addAttribute("body", post1.getBody());
-        model.addAttribute("post", postDao.getOne(id));
+    public String getPost(@PathVariable long id, Model model, Principal principal){
+        String userName = "";
+        if (principal != null) {
+            userName = principal.getName();
+        }
+        model.addAttribute("userName", userName);
+        model.addAttribute("post",postDao.getOne(id));
         return "posts/show";
     }
 
     @GetMapping("/posts/create")
-    public String getCreatePostForm(){
-    Post newPost = new Post();
-    newPost.setTitle("New Post");
-    newPost.setBody("This is the newly saved post");
-    newPost.setUser(userDao.getOne(1L));
-    postDao.save(newPost);
-    String emailSubject = "This is the email subject";
-    String emailBody = " This is the body of the email";
-    emailService.prepareandsend(newPost, emailSubject, emailBody);
+    public String getCreatePostForm(Model view){
+        view.addAttribute("action", "create");
+        view.addAttribute("post", new Post());
         return "posts/create";
     }
 
+
     @PostMapping("/posts/create")
-    public String createPost(@ModelAttribute Post post){
-        post.setUser(userDao.getOne(1l));
+    public String getCreatePostForm(@ModelAttribute Post post){
+        User loggedInUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        post.setUser(loggedInUser);
         postDao.save(post);
+        emailService.prepareandsend(post, "Post Created!", "We created your post: " + post.getTitle());
         return "redirect:/posts";
     }
 
-
-    @RequestMapping(path="/posts", method=RequestMethod.DELETE)
+    @PostMapping("/posts/{id}/delete")
     public String delete(@PathVariable long id){
-        postDao.deleteById(id);
-        return "redirect:posts";
+       User loggedInUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+       if (loggedInUser.getId() == postDao.getOne(id).getUser().getId())
+       postDao.deleteById(id);
+        return "redirect:/posts";
     }
 
     @GetMapping("/posts/{id}/edit")
